Fix `Floki.raw_html/1` for `script` and `style` tags
It does not try to encode HTML entities inside
script or style tags.

It is related with: https://github.com/philss/floki/issues/161
