+++ lib/floki.ex
@@ -72,25 +72,6 @@ defmodule Floki do
     HTMLParser.parse(html)
   end
 
-  @self_closing_tags [
-    "area",
-    "base",
-    "br",
-    "col",
-    "command",
-    "embed",
-    "hr",
-    "img",
-    "input",
-    "keygen",
-    "link",
-    "meta",
-    "param",
-    "source",
-    "track",
-    "wbr"
-  ]
-
   @doc """
   Converts HTML tree to raw HTML.
   Note that the resultant HTML may be different from the original one.
@@ -105,59 +86,7 @@ defmodule Floki do
 
   @spec raw_html(html_tree | binary) :: binary
 
-  def raw_html(html_tree), do: raw_html(html_tree, "")
-  defp raw_html([], html), do: html
-  defp raw_html(string, _html) when is_binary(string), do: HtmlEntities.encode(string)
-  defp raw_html(tuple, html) when is_tuple(tuple), do: raw_html([tuple], html)
-
-  defp raw_html([string | tail], html) when is_binary(string) do
-    raw_html(tail, html <> HtmlEntities.encode(string))
-  end
-
-  defp raw_html([{:comment, comment} | tail], html),
-    do: raw_html(tail, html <> "<!--#{comment}-->")
-
-  defp raw_html([{:pi, "xml", attrs} | tail], html) do
-    raw_html(tail, html <> "<?xml " <> tag_attrs(attrs) <> "?>")
-  end
-
-  defp raw_html([{:doctype, type, public, system} | tail], html) do
-    attr =
-      case {public, system} do
-        {"", ""} -> ""
-        {"", system} -> " SYSTEM \"#{system}\""
-        {public, system} -> " PUBLIC \"#{public}\" \"#{system}\""
-      end
-
-    raw_html(tail, html <> "<!DOCTYPE #{type}#{attr}>")
-  end
-
-  defp raw_html([{type, attrs, children} | tail], html) do
-    raw_html(tail, html <> tag_for(type, tag_attrs(attrs), children))
-  end
-
-  defp tag_attrs(attr_list) do
-    attr_list
-    |> Enum.reduce("", &build_attrs/2)
-    |> String.trim()
-  end
-
-  defp build_attrs({attr, value}, attrs), do: ~s(#{attrs} #{attr}="#{value}")
-  defp build_attrs(attr, attrs), do: "#{attrs} #{attr}"
-
-  defp tag_for(type, attrs, []) when type in @self_closing_tags do
-    case attrs do
-      "" -> "<#{type}/>"
-      _ -> "<#{type} #{attrs}/>"
-    end
-  end
-
-  defp tag_for(type, attrs, children) do
-    case attrs do
-      "" -> "<#{type}>#{raw_html(children)}</#{type}>"
-      _ -> "<#{type} #{attrs}>#{raw_html(children)}</#{type}>"
-    end
-  end
+  defdelegate raw_html(html_tree), to: Floki.RawHTML
 
   @doc """
   Find elements inside a HTML tree or string.
